Refactor code formatting and add User model
This commit mainly focuses on enhancing code readability and cleanliness. Code formatting was improved in multiple files for better consistency. For instance, missing semicolons were added, and the standard for quotation marks was unified. Besides this, a new model 'User' was introduced in 'user.model.ts' to represent user data class more effectively. The tsconfig.json file was also updated to maintain a streamlined format for the "exclude" array.